Algorithm/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask~=3.0.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy~=1.24.4
opencv-python~=3.4.10.35
openvino==2022.1.0
requests==2.28.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0

pillow~=9.2.0
torch~=1.10.1+cu113
torchvision~=0.11.2+cu113